requirements.txt
absl-py==1.4.0
aiobotocore==2.16.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
asn1crypto==1.5.1
asttokens==3.0.0
atpublic==5.0
attrs==24.2.0
backports.tarfile==1.2.0
bcrypt==4.2.1
blinker==1.9.0
botocore==1.35.88
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
extra-streamlit-components==0.1.71
filelock==3.16.1
frozenlist==1.5.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.31.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
langchain==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.4
langsmith==0.2.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
narwhals==1.17.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytimeparse==1.1.8
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
s3fs==2023.12.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snowflake==1.0.2
snowflake-connector-python==3.12.4
snowflake-ml-python==1.7.2
snowflake-snowpark-python==1.26.0
snowflake._legacy==1.0.0
snowflake.core==1.0.2
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.41.0
streamlit-chat==0.1.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.0
xgboost==2.1.3
yarl==1.18.3
zipp==3.21.0
